// src/app/enums/demand-status.enum.ts
export enum DemandStatusEnum {
  RASCUNHO = "RASCUNHO",
  CADASTRADO_VISTORIA = "CADASTRADO_VISTORIA",
  ESPERANDO_MELHORIA = "ESPERANDO_MELHORIA",
  EM_ANALISE = "EM_ANALISE",
  ESPERANDO_OBRA = "ESPERANDO_OBRA",
  CONCLUIR_OBRAS = "CONCLUIR_OBRAS",
  ESPERANDO_VALIDACAO = "ESPERANDO_VALIDACAO",
  CONCLUIDO = "CONCLUIDO",
  RESCISAO = "RESCISAO",
  CANCELADO = "CANCELADO",
}